requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langgraph==0.5.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.73
langsmith==0.4.7
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.1
openai==1.97.0
orjson==3.11.0
ormsgpack==1.10.0
packaging==25.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
